refactor: hoist doc ops into engine::ops layer, dedupe store wiring
Move create/link/delete/update/fix/resolve op cores from CLI into a new
engine::ops module with no clap types in signatures. TUI no longer imports
crate::cli (AC1); CLI and TUI are thin callers of the shared engine layer.

Dedupe the ClickUp write-store token+store construction into a single
clickup_write_store helper in store_dispatch, called from create/update/
delete ops and the TUI edit path (AC3). build_registry read store (token
None) unchanged.

Human fix formatting stays frontend-side: format_human back in cli::fix,
TUI uses a local formatter over the engine FixOutput type.

Implements STORY-212 (AC1-AC4). Existing tests pass unmodified.
