+++ src/cli/fix/output.rs
@@ -1,38 +1,4 @@
-use super::{ConfigFixResult, FixOutput};
-
-pub(super) fn format_config_human(result: &ConfigFixResult, dry_run: bool) -> String {
-    let mut out = String::new();
-
-    for name in &result.relationships_added {
-        if dry_run {
-            out.push_str(&format!("Would add relationship {}\n", name));
-        } else {
-            out.push_str(&format!("Added relationship {}\n", name));
-        }
-    }
-
-    for name in &result.rules_added {
-        if dry_run {
-            out.push_str(&format!("Would add rule {}\n", name));
-        } else {
-            out.push_str(&format!("Added rule {}\n", name));
-        }
-    }
-
-    for name in &result.lifecycles_added {
-        if dry_run {
-            out.push_str(&format!("Would add default lifecycle to type {}\n", name));
-        } else {
-            out.push_str(&format!("Added default lifecycle to type {}\n", name));
-        }
-    }
-
-    if out.is_empty() {
-        out.push_str("Config already up to date; nothing to add\n");
-    }
-
-    out
-}
+use crate::engine::ops::fix::{ConfigFixResult, FixOutput};
 
 pub(super) fn format_human(output: &FixOutput, dry_run: bool) -> String {
     let mut result = String::new();
@@ -87,4 +53,38 @@ pub(super) fn format_human(output: &FixOutput, dry_run: bool) -> String {
     }
 
     result
+}
+
+pub(super) fn format_config_human(result: &ConfigFixResult, dry_run: bool) -> String {
+    let mut out = String::new();
+
+    for name in &result.relationships_added {
+        if dry_run {
+            out.push_str(&format!("Would add relationship {}\n", name));
+        } else {
+            out.push_str(&format!("Added relationship {}\n", name));
+        }
+    }
+
+    for name in &result.rules_added {
+        if dry_run {
+            out.push_str(&format!("Would add rule {}\n", name));
+        } else {
+            out.push_str(&format!("Added rule {}\n", name));
+        }
+    }
+
+    for name in &result.lifecycles_added {
+        if dry_run {
+            out.push_str(&format!("Would add default lifecycle to type {}\n", name));
+        } else {
+            out.push_str(&format!("Added default lifecycle to type {}\n", name));
+        }
+    }
+
+    if out.is_empty() {
+        out.push_str("Config already up to date; nothing to add\n");
+    }
+
+    out
 }
