+++ src/engine/ops.rs
@@ -1,0 +1,9 @@
+//! Operation layer shared by the CLI and TUI: create/link/delete/update/fix
+//! cores, free of argument parsing and output formatting.
+
+pub mod create;
+pub mod delete;
+pub mod fix;
+pub mod link;
+pub mod resolve;
+pub mod update;
